requirements.txt
python-telegram-bot==20.0
gspread
oauth2client
nest_asyncio